:bricks: move member modal into its own component
