+++ next/components/small_navbar/SmallNavbar.modules.css
@@ -1,0 +1,34 @@
+.header {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 100%;
+  align-items: center;
+  padding: 0.25rem var(--gutter);
+  position: absolute;
+  z-index: 1000;
+  isolation: isolate;
+  height: 8.5rem;
+}
+
+.nav {
+  display: flex;
+  align-items: center;
+  gap: 2.5rem;
+}
+
+.nav a {
+  font-size: 0.9rem;
+  transition: color 0.4s ease-out;
+  letter-spacing: 1px;
+}
+
+@media (max-width: 800px) {
+  .header {
+    flex-direction: column;
+  }
+  .youniteLogo {
+    position: relative;
+  }
+}
+
